requirements.txt
python_telegram_bot==13.7
apscheduler==3.7.0